requirements.txt

# Core Utilities

pandas==2.3.1
numpy==2.3.1
scikit-learn==1.7.0
scipy==1.16.0
tqdm==4.66.4
joblib==1.4.2


# NLP & Text Processing

nltk==3.9.1
spacy==3.8.7
# Download English model automatically:
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
rapidfuzz==3.9.3  # for fast fuzzy matching of skills


# Embeddings and Transformers

sentence-transformers==3.0.1
transformers==4.57.1
torch==2.5.1


# PDF & Text Extraction

pdfminer.six==20221105
PyPDF2==3.0.1
python-docx==1.1.2


# Web App

streamlit==1.51.0


# Visualization & Debugging

plotly==6.2.0
seaborn==0.13.2
matplotlib==3.10.3